Reorganize Landing Page Components into Dedicated Page Folder
