+++ web-app/src/app/(landing)/page/hero.tsx
