+++ web-app/src/app/(landing)/page/how-it-works.tsx
