+++ web-app/src/app/(landing)/page/join-our-community.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import { useTranslations } from "next-intl";
 
-import { DiscordButton, XButton } from "./components/social-button";
+import { DiscordButton, XButton } from "../components/social-button";
 
 export function JoinOurCommunity() {
   const t = useTranslations("Landing.JoinOurCommunity");
